+++ examples/data-sources/iosxe_interface_ospf_process/data-source.tf
@@ -1,0 +1,5 @@
+data "iosxe_interface_ospf_process" "example" {
+  type       = "GigabitEthernet"
+  name       = "2"
+  process_id = 1
+}