+++ examples/resources/iosxe_interface_ospf_process/resource.tf
@@ -1,0 +1,10 @@
+resource "iosxe_interface_ospf_process" "example" {
+  type       = "GigabitEthernet"
+  name       = "2"
+  process_id = 1
+  area = [
+    {
+      area_id = "1"
+    }
+  ]
+}